+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - A components page on the website.
 - Leptos Icons with Font Awesome package on the website.
 - More props to modal Leptos component and included close button by default.
+- Highlight.js for code examples.
 
 ### Removed